fix(settings): group Korean news sources
